Run the diem node in a separate process
